requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
attrs==19.3.0
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.8
entrypoints==0.3
feedparser==5.2.1
Flask==1.0.3
future==0.18.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
pathspec==0.7.0
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
python-telegram-bot==11.1.0
pytz==2019.1
regex==2020.1.8
requests==2.22.0
rsa==4.0
six==1.12.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.3
Werkzeug==0.15.4
